refactor(synthesize): generator client + PromptFrames kernel — the god-file pass
synthesize.service.ts had crept to 916 physical lines (past the
800-line gate spirit, surviving on skipComments) and the V10 points
were drilling three frame booleans through two layers.

1. runGenerator moves to generator-client.ts at the same seam as
   verifier.ts — a self-contained LLM call owning system-prompt
   selection, the strict-JSON schema, the W5 #24 truncation salvage,
   usage accounting and tracing; the orchestrator keeps a thin
   callGenerator adapter (916 → 762 lines).

2. resolvePromptFrames (evidence-gates) is the single flag→frame
   mapping: orderingFrame (needs the fired record), the
   date-arbitrated conflict frame (rides updateStoryRendering) and
   the topic-record header (query_arc) resolve in one tested kernel
   instead of inline at the call site.

Behavior-neutral: 236 suites / 1986 tests, tsc, lint green.
