+++ src/synthesize/synthesize.service.ts
@@ -5,7 +5,6 @@ import { createOpenAiClientOrThrow } from '../ai/openai-client';
 import { SearchService, SearchHit } from '../search/search.service';
 import { Semaphore } from '../common/semaphore';
 import { withSpan } from '../common/tracing';
-import { withGenAiCall } from '../common/gen-ai-observability';
 import { clampLlmInputText } from '../common/input-limits';
 import { traceArtifact } from '../common/debug-trace';
 import { getAbortSignal } from '../common/request-context';
@@ -22,10 +21,12 @@ import {
   resolveAnswerLang,
   resolveCitations,
   resolveLaneDateContext,
-  salvageTruncatedAnswer,
   verifierErrorResult,
 } from './synthesize.helpers';
-import { wantsTimelineEvidence } from './evidence-gates';
+import {
+  resolvePromptFrames,
+  wantsTimelineEvidence,
+} from './evidence-gates';
 import { applyEvidenceUnion } from './evidence-union';
 import {
   routeLane,
@@ -42,7 +43,8 @@ import { buildFactIndex } from './fact-index';
 import { appendUpdateStories } from './update-story';
 export { buildFactIndex } from './fact-index';
 import type { Citation } from './fact-index';
-import { buildGeneratorUserMessage } from './generator-prompt';
+import { runGenerator } from './generator-client';
+import type { GenerateRequest } from './generator-client';
 import { runVerifier, type VerifierOutput } from './verifier';
 export { buildGeneratorUserMessage } from './generator-prompt';
 import type { SearchDto } from '../search/dto/search.dto';
@@ -87,30 +89,6 @@ import type {
   SynthesisReason,
   SynthesizeResult,
 } from './synthesize.types';
-
-const GENERATOR_SYSTEM = `You are an answer synthesizer for a knowledge graph.
-
-Given a user query and a set of retrieved facts (each prefixed with its factId in square brackets, e.g. "[knowledge_fact:8a3fd2c1b9e4f7a6d5c0] ..."), generate a CONCISE answer that:
-1. Uses ONLY information present in the provided facts. Do NOT speculate, fill in missing details, or use outside knowledge.
-2. After each claim in the answer, inline a citation in square brackets, copying the factId EXACTLY as it appears in the fact list — including its "knowledge_fact:" prefix. Do not abbreviate, renumber, or change the prefix. Example: "Maya complained about a broken washing machine [knowledge_fact:8a3fd2c1b9e4f7a6d5c0]". Mirror every factId you cite inline into the citedFactIds array.
-3. If the facts do not answer the question, output the exact answer string "I don't have grounded evidence for that." with citedFactIds set to [].
-
-Output strictly the JSON shape requested by the schema. Do not include preamble, follow-ups, or chain-of-thought.`;
-
-/**
- * Best-effort / never-abstain generator (guardrails='answer'). Same grounding
- * discipline, but instead of refusing when the facts are thin it commits to
- * the single most likely SHORT answer the retrieved facts point to. For QA
- * settings where an abstention scores strictly worse than a best guess.
- */
-const GENERATOR_SYSTEM_ANSWER = `You are an answer synthesizer for a knowledge graph.
-
-Given a user query and a set of retrieved facts (each prefixed with its factId in square brackets, e.g. "[knowledge_fact:8a3fd2c1b9e4f7a6d5c0] ..."), generate a SHORT, CONCRETE answer that:
-1. Is grounded in the provided facts — prefer specifics stated in them; do not invent named entities, dates, or numbers that no fact supports.
-2. After each claim, inline a citation copying the factId EXACTLY (including its "knowledge_fact:" prefix), and mirror every cited factId into citedFactIds.
-3. ALWAYS commit to an answer. If the facts do not fully resolve the question, give the single most likely short answer they point to — do NOT refuse, do NOT output "I don't have grounded evidence", do NOT hedge with "the facts don't say". Answer in as few words as the question allows.
-
-Output strictly the JSON shape requested by the schema. Do not include preamble, follow-ups, or chain-of-thought.`;
 
 /**
  * SynthesizeService — orchestrates the corrective-RAG flow:
@@ -336,15 +314,9 @@ export class SynthesizeService {
               transcriptLines,
               insightLines,
               timelineEvidence,
-              // V10 §3: dedicated order-of-mention frame — only when
-              // the mention record actually fired for this query.
-              orderingFrame: profile.orderingFrame && timelineEvidence,
-              // V10 §4: the insight slot carries a query-time topic
-              // record — the header must say so.
-              arcInsights: profile.insightEvidence === 'query_arc',
-              // V10 §2b: date-arbitrated conflict frame rides the
-              // update-story flag (one lifecycle-read story).
-              dateArbitratedConflicts: profile.updateStoryRendering,
+              // V10 frame switches — resolved once by the kernel
+              // (evidence-gates.resolvePromptFrames), not inline.
+              ...resolvePromptFrames(profile, timelineEvidence),
               model,
               answerLang,
               neverAbstain: guardrails === 'answer',
@@ -752,149 +724,20 @@ export class SynthesizeService {
     return { results, factIndex, factLines };
   }
 
-  private async callGenerator({
-    query,
-    factLines,
-    transcriptLines,
-    insightLines,
-    timelineEvidence,
-    orderingFrame,
-    dateArbitratedConflicts,
-    arcInsights,
-    model,
-    answerLang,
-    neverAbstain = false,
-    dateContext,
-    lane,
-    instructions,
-    conflicts,
-  }: {
-    query: string;
-    factLines: string[];
-    /** Episodic-lane quotes (P2) — rendered as a separate typed section. */
-    transcriptLines?: string[];
-    /** V8 §1: derived insights — their own section, own budget slot. */
-    insightLines?: string[];
-    /** V8 §2: transcript excerpts are the mention record (ordering ask). */
-    timelineEvidence?: boolean;
-    /** V10 §3: order-of-mention frame replaces the enumeration frame. */
-    orderingFrame?: boolean;
-    /** V10 §2b: date-arbitrated conflict frame (updateStoryRendering). */
-    dateArbitratedConflicts?: boolean;
-    /** V10 §4: insight lines are a query-time topic record. */
-    arcInsights?: boolean;
-    model: string;
-    answerLang: string | null;
-    neverAbstain?: boolean;
-    /** ISO date the answer should treat as "today" (SYNTHESIZE_DATE_CONTEXT). */
-    dateContext?: string;
-    /** T1 typed dispatch lane, when the router matched. */
-    lane?: LaneId | null;
-    /** T7: standing user instructions for their own section. */
-    instructions?: string[];
-    /** T3: COMPETING conflict pairs detected in the evidence. */
-    conflicts?: Array<{ factIds: string[]; label: string }>;
-  }): Promise<GeneratorOutput> {
-    const systemPrompt = neverAbstain
-      ? GENERATOR_SYSTEM_ANSWER
-      : GENERATOR_SYSTEM;
-    const user = buildGeneratorUserMessage({
-      query,
-      factLines,
-      transcriptLines,
-      insightLines,
-      timelineEvidence,
-      orderingFrame,
-      dateArbitratedConflicts,
-      arcInsights,
-      answerLang,
-      dateContext,
-      lane,
-      instructions,
-      conflicts,
+  /**
+   * Thin adapter over the generator client (V10 architecture pass) —
+   * same seam as callVerifier/runVerifier: the orchestrator supplies
+   * its client/metrics/logger, the module owns the call.
+   */
+  private async callGenerator(
+    args: Omit<GenerateRequest, 'openai' | 'metrics' | 'logger'>,
+  ): Promise<GeneratorOutput> {
+    return runGenerator({
+      openai: this.openai,
+      metrics: this.metrics,
+      logger: this.logger,
+      ...args,
     });
-    traceArtifact('synthesize.generator_prompt', {
-      system: systemPrompt,
-      user,
-      model,
-      answerLang,
-    });
-    const res = await withGenAiCall(
-      {
-        kind: 'chat',
-        spanName: 'gen_ai.chat.synthesize_generator',
-        system: 'openai',
-        model,
-        attrs: { 'brain.synthesize.answer_lang': answerLang ?? 'auto' },
-      },
-      this.metrics,
-      () => this.openai.chat.completions.create(
-      {
-      model,
-      messages: [
-        { role: 'system', content: systemPrompt },
-        { role: 'user', content: user },
-      ],
-      response_format: {
-        type: 'json_schema',
-        json_schema: {
-          name: 'synthesized_answer',
-          strict: true,
-          schema: {
-            type: 'object',
-            additionalProperties: false,
-            properties: {
-              answer: { type: 'string' },
-              citedFactIds: { type: 'array', items: { type: 'string' } },
-            },
-            required: ['answer', 'citedFactIds'],
-          },
-        },
-      },
-      max_completion_tokens: 512,
-      temperature: 0,
-    }, { signal: getAbortSignal() }),
-    );
-    const content = res.choices[0]?.message?.content;
-    if (!content) throw new Error('empty generator response');
-    // Audit W5 #24: the exhaustive-list lanes ("a partial list is a wrong
-    // answer") run against a 512-token cap. A truncated strict-JSON body
-    // used to throw here and degrade to `generator_error` — i.e. the
-    // lanes that must enumerate silently ABSTAINED instead of returning
-    // what they had. Say so explicitly in the trace, and salvage the
-    // answer text when the JSON envelope is the only casualty.
-    const finishReason = res.choices[0]?.finish_reason;
-    let parsed: GeneratorOutput;
-    try {
-      parsed = JSON.parse(content) as GeneratorOutput;
-    } catch (err) {
-      if (finishReason !== 'length') throw err;
-      const salvaged = salvageTruncatedAnswer(content);
-      if (!salvaged) throw err;
-      this.logger.warn(
-        'generator output hit the token cap; salvaged the partial answer',
-      );
-      this.metrics?.countSynthesize('generator_truncated');
-      traceArtifact('synthesize.generator_truncated', {
-        finishReason,
-        salvagedChars: salvaged.answer.length,
-      });
-      parsed = salvaged;
-    }
-    if (typeof parsed.answer !== 'string') {
-      throw new Error('generator returned non-string answer');
-    }
-    if (!Array.isArray(parsed.citedFactIds)) {
-      parsed.citedFactIds = [];
-    }
-    if (res.usage) {
-      parsed.usage = {
-        promptTokens: res.usage.prompt_tokens ?? 0,
-        completionTokens: res.usage.completion_tokens ?? 0,
-      };
-    }
-    traceArtifact('synthesize.generator_output', parsed);
-    return parsed;
   }
 
   private async callVerifier(args: {
